Refactor ray casting and expose to python
Remove duplicated Ray3D class in favor of QgsRay3D, add missing methods.
Remove qgsraycastingutils_p.h, keep ray intersection code in qgsraycastingutils.h/.cpp
Introduce QgsRayCastContext class for ray casting parameters, including maxDistance which was once part of Ray3D,
and make available in python bindings.
Introduce QgsRayCastHit to encapsulate individual ray hits and QgsRayCastResult to handle all the hits of a
ray casting operation and make available in python bindings.
Introduce Qgs3DMapCanvas::castRay() and make available in python.

+++ src/3d/qgs3dmapscene.h
@@ -132,7 +132,7 @@ class _3D_EXPORT Qgs3DMapScene : public QObject
      * Given screen error (in pixels) and distance from camera (in 3D world coordinates), this function
      * estimates the error in world space. Takes into account camera's field of view and the screen (3D view) size.
      */
-    float worldSpaceError( float epsilon, float distance ) const;
+    double worldSpaceError( double epsilon, double distance ) const;
 
     /**
      * Exports the scene according to the scene export settings
